// src/app/core/models/my-profile.model.ts
export class MyProfile {
    firt_name: string;
    last_name: string;
    preferred_name: string;
    email: string;
    password: string;
    phone_dial_code: string;
    phone_number: string;
    address_line1: string;
    address_line2: string;
    address_line3: string;
    postal_code: string;
    state_code: string;
    country_code: string;
    my_advisor_id: string;
    my_profile_photo_key: string;
    salt: string;
    advisor_name: string;
    recruiter_name: string;
    teamLeader_name: string;
    branchManager_name: string;
    advisor_id: string;
    recruiter_id: string;
    teamLeader_id: string;
    branchManager_id: string;
    advisor_photo_key: string;
    recruiter_photo_key: string;
    teamLeader_photo_key: string;
    branchManager_photo_key: string;
    uuid: string;
    is_active: boolean;
    is_registered: boolean;
    is_anomynous_account : boolean;
    // constructor(
    //     firstName: string,
    //     lastName: string,
    //     preferredname : string,
    //     email: string,
    //     password: string,
    //     phone_dial_code: string,
    //     phone_number: string,
    //     postal_code : string,
    //     city_state : string,
    //     detail_address : string,
    //     unit_number : string ,
    //     country_code : string,
    //     my_advisor_id: string,
    //     salt: string,
    //     advisor_name: string,
    //     recruiter_name: string,
    //     teamLeader_name: string,
    //     branchManager_name: string,
    //     advisor_id: string,
    //     recruiter_id: string,
    //     teamLeader_id: string,
    //     branchManager_id: string,)
    // {
    //     this.firt_name = firstName;
    //     this.last_name = lastName;
    //     this.preferred_name = preferredname;
    //     this.email = email;
    //     this.password = password;
    //     this.phone_dial_code = phone_dial_code;
    //     this.phone_number = phone_number;
    //     this.my_advisor_id = my_advisor_id;
    //     this.salt=salt;
    //     this.advisor_name=advisor_name;
    //     this.recruiter_name=recruiter_name;
    //     this.teamLeader_name=teamLeader_name;
    //     this.branchManager_name=branchManager_name;
    //     this.advisor_id=advisor_id;
    //     this.recruiter_id=recruiter_id;
    //     this.teamLeader_id=teamLeader_id;
    //     this.branchManager_id=branchManager_id;
    //     this.postal_code=postal_code;
    //     this.city_state=city_state;
    //     this.unit_number=unit_number;
    //     this.detail_address=detail_address;
    //     this.country_code=country_code;
    // }
}